requirements.txt
discord.py==2.3.1
rich==13.5.2
